pyproject.toml
[tool.poetry]
name = "camcal"
version = "0.1.0"
description = "Small helper package to perform camera orientation calibration and automated processing of images taken"
authors = ["nfb2021 <nicolas.bader@geo.tuwien.ac.at>"]
readme = "README.md"
packages = [{include = "camcal", from = "camcal/src"}]

[tool.poetry.dependencies]
python = "^3.13"
opencv-python-headless = "^4.10.0.84"
opencv-contrib-python = "^4.10.0.84"
pydantic = "^2.0"
ipykernel = "^6.29.5"
numpy = "^2.1.2"
screeninfo = "^0.8.1"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.2"
scipy = "^1.14.1"
polars = "^1.12.0"
tqdm = "^4.67.0"
xarray = "2024.07.0"
natsort = "^8.4.0"
dask = "^2024.11.1"
netcdf4 = "^1.7.2"
pillow = "^11.0.0"
exifread = "^3.0.0"
xarray-datatree = "^0.0.15"
zarr = "^2.18.3"
pint = "^0.24.4"
notebook = "^7.2.2"
ipympl = "^0.9.4"
ipywidgets = "^8.1.5"
pylint = "^3.3.1"
pylint-pydantic = "^0.3.2"
flake8 = "^7.1.1"
flake8-pydantic = "^0.3.1"
mypy = "^1.13.0"
astral = "^3.2"
pytz = "^2024.2"
types-pytz = "^2024.2.0.20241003"
types-exifread = "^3.0.0.20240806"
types-tqdm = "^4.67.0.20241119"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10.0"
mypy = "^1.13.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
coverage = "^7.6.4"
pre-commit = "^3.5.0"  # Added pre-commit hooks for consistent code formatting
pytest-cov = "^4.1.0"  # For pytest with coverage integration

[tool.mypy]
plugins = ["pydantic.mypy"]  # Ensures pydantic-specific type checking
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.pylint]
load-plugins = ["pylint_pydantic"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.isort]
profile = "black"

[tool.coverage.run]
branch = true
source = ["camcal"]

[tool.pre-commit]
repos = [
    {repo = "https://github.com/pre-commit/pre-commit-hooks", rev = "v4.4.0", hooks = [{id = "end-of-file-fixer"}, {id = "trailing-whitespace"}]},
    {repo = "https://github.com/psf/black", rev = "23.9.1", hooks = [{id = "black"}]},
    {repo = "https://github.com/pre-commit/mirrors-isort", rev = "v5.12.0", hooks = [{id = "isort"}]},
    {repo = "https://github.com/pre-commit/mirrors-flake8", rev = "v7.1.1", hooks = [{id = "flake8"}]}
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
